+++ HISTORY.md
@@ -16,6 +16,7 @@ http://visjs.org
         }
 - Fixed the height of background and foreground panels of groups.
 - Fixed ranges in the Timeline sometimes overlapping when dragging the Timeline.
+- Fixed `DataView` not working in Timeline.
 
 ### Network